+++ plugins/catalog-react/src/alpha/blueprints/EntityCardBlueprint.test.tsx
@@ -51,15 +51,134 @@ describe('EntityCardBlueprint', () => {
             "$schema": "http://json-schema.org/draft-07/schema#",
             "additionalProperties": false,
             "properties": {
-              "area": {
-                "enum": [
-                  "peek",
-                  "info",
-                  "full",
-                ],
-                "type": "string",
-              },
               "filter": {
+                "anyOf": [
+                  {
+                    "type": "string",
+                  },
+                  {
+                    "anyOf": [
+                      {
+                        "anyOf": [
+                          {
+                            "type": [
+                              "string",
+                              "number",
+                              "boolean",
+                            ],
+                          },
+                          {
+                            "items": {
+                              "$ref": "#/properties/filter/anyOf/1/anyOf/0/anyOf/0",
+                            },
+                            "type": "array",
+                          },
+                        ],
+                      },
+                      {
+                        "additionalProperties": false,
+                        "properties": {
+                          "$all": {
+                            "items": {
+                              "$ref": "#/properties/filter/anyOf/1",
+                            },
+                            "type": "array",
+                          },
+                        },
+                        "required": [
+                          "$all",
+                        ],
+                        "type": "object",
+                      },
+                      {
+                        "additionalProperties": false,
+                        "properties": {
+                          "$any": {
+                            "items": {
+                              "$ref": "#/properties/filter/anyOf/1",
+                            },
+                            "type": "array",
+                          },
+                        },
+                        "required": [
+                          "$any",
+                        ],
+                        "type": "object",
+                      },
+                      {
+                        "additionalProperties": false,
+                        "properties": {
+                          "$not": {
+                            "$ref": "#/properties/filter/anyOf/1",
+                          },
+                        },
+                        "required": [
+                          "$not",
+                        ],
+                        "type": "object",
+                      },
+                      {
+                        "additionalProperties": {
+                          "anyOf": [
+                            {
+                              "$ref": "#/properties/filter/anyOf/1/anyOf/0",
+                            },
+                            {
+                              "additionalProperties": false,
+                              "properties": {
+                                "$exists": {
+                                  "type": "boolean",
+                                },
+                              },
+                              "required": [
+                                "$exists",
+                              ],
+                              "type": "object",
+                            },
+                            {
+                              "additionalProperties": false,
+                              "properties": {
+                                "$in": {
+                                  "items": {
+                                    "$ref": "#/properties/filter/anyOf/1/anyOf/0/anyOf/0",
+                                  },
+                                  "type": "array",
+                                },
+                              },
+                              "required": [
+                                "$in",
+                              ],
+                              "type": "object",
+                            },
+                            {
+                              "additionalProperties": false,
+                              "properties": {
+                                "$contains": {
+                                  "$ref": "#/properties/filter/anyOf/1",
+                                },
+                              },
+                              "required": [
+                                "$contains",
+                              ],
+                              "type": "object",
+                            },
+                          ],
+                        },
+                        "propertyNames": {
+                          "pattern": "^(?!\\$).*$",
+                        },
+                        "type": "object",
+                      },
+                    ],
+                  },
+                ],
+              },
+              "type": {
+                "enum": [
+                  "summary",
+                  "info",
+                  "content",
+                ],
                 "type": "string",
               },
             },
@@ -96,7 +215,7 @@ describe('EntityCardBlueprint', () => {
             "config": {
               "optional": true,
             },
-            "id": "catalog.entity-card-area",
+            "id": "catalog.entity-card-type",
             "optional": [Function],
             "toString": [Function],
           },