+++ plugins/catalog-react/src/alpha/blueprints/EntityCardBlueprint.ts
@@ -22,9 +22,14 @@ import {
 import {
   entityFilterFunctionDataRef,
   entityFilterExpressionDataRef,
-  entityCardAreaDataRef,
-  defaultEntityCardAreas,
+  entityCardTypeDataRef,
+  entityCardTypes,
+  EntityCardType,
 } from './extensionData';
+import { createEntityPredicateSchema } from '../predicates/createEntityPredicateSchema';
+import { EntityPredicate } from '../predicates';
+import { resolveEntityFilterData } from './resolveEntityFilterData';
+import { Entity } from '@backstage/catalog-model';
 
 /**
  * @alpha
@@ -37,50 +42,43 @@ export const EntityCardBlueprint = createExtensionBlueprint({
     coreExtensionData.reactElement,
     entityFilterFunctionDataRef.optional(),
     entityFilterExpressionDataRef.optional(),
-    entityCardAreaDataRef.optional(),
+    entityCardTypeDataRef.optional(),
   ],
   dataRefs: {
     filterFunction: entityFilterFunctionDataRef,
     filterExpression: entityFilterExpressionDataRef,
-    area: entityCardAreaDataRef,
+    type: entityCardTypeDataRef,
   },
   config: {
     schema: {
-      filter: z => z.string().optional(),
-      area: z => z.enum(defaultEntityCardAreas).optional(),
+      filter: z =>
+        z.union([z.string(), createEntityPredicateSchema(z)]).optional(),
+      type: z => z.enum(entityCardTypes).optional(),
     },
   },
   *factory(
     {
       loader,
       filter,
-      defaultArea,
+      type,
     }: {
       loader: () => Promise<JSX.Element>;
-      filter?:
-        | typeof entityFilterFunctionDataRef.T
-        | typeof entityFilterExpressionDataRef.T;
-      defaultArea?: (typeof defaultEntityCardAreas)[number];
+      filter?: string | EntityPredicate | ((entity: Entity) => boolean);
+      type?: EntityCardType;
     },
     { node, config },
   ) {
     yield coreExtensionData.reactElement(ExtensionBoundary.lazy(node, loader));
 
-    if (config.filter) {
-      yield entityFilterExpressionDataRef(config.filter);
-    } else if (typeof filter === 'string') {
-      yield entityFilterExpressionDataRef(filter);
-    } else if (typeof filter === 'function') {
-      yield entityFilterFunctionDataRef(filter);
-    }
+    yield* resolveEntityFilterData(filter, config, node);
 
-    const area = config.area ?? defaultArea;
-    if (area) {
-      yield entityCardAreaDataRef(area);
+    const finalType = config.type ?? type;
+    if (finalType) {
+      yield entityCardTypeDataRef(finalType);
     } else {
       // eslint-disable-next-line no-console
       console.warn(
-        `DEPRECATION WARNING: Not providing defaultArea for entity cards is deprecated. Missing from '${node.spec.id}'`,
+        `DEPRECATION WARNING: Not providing type for entity cards is deprecated. Missing from '${node.spec.id}'`,
       );
     }
   },