+++ plugins/catalog-react/src/alpha/blueprints/extensionData.tsx
@@ -45,21 +45,25 @@ export const defaultEntityContentGroups = {
 };
 
 /** @internal */
-export const entityContentGroupDataRef = createExtensionDataRef<
-  false | string
->().with({
+export const entityContentGroupDataRef = createExtensionDataRef<string>().with({
   id: 'catalog.entity-content-group',
 });
 
 /**
- * @alpha
- * Default entity content groups.
+ * @internal
+ * Available entity card types
  */
-export const defaultEntityCardAreas = ['peek', 'info', 'full'] as const;
+export const entityCardTypes = [
+  'summary',
+  'info',
+  'content',
+] as const satisfies readonly EntityCardType[];
+
+/** @alpha */
+export type EntityCardType = 'summary' | 'info' | 'content';
 
 /** @internal */
-export const entityCardAreaDataRef = createExtensionDataRef<
-  (typeof defaultEntityCardAreas)[number]
->().with({
-  id: 'catalog.entity-card-area',
-});
+export const entityCardTypeDataRef =
+  createExtensionDataRef<EntityCardType>().with({
+    id: 'catalog.entity-card-type',
+  });